refactor: extract mount request tool owner
