+++ crates/agent-engine/src/runtime/virtual_tool.rs
@@ -1,0 +1,7 @@
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(super) enum VirtualToolOutcome {
+    NotVirtual,
+    Continue { refresh_context: bool },
+    PauseTurn,
+    EndTurn,
+}
